+++ test/window_test.cpp
@@ -3,7 +3,6 @@
 #include <cstdlib>
 #include <chrono>
 #include <thread>
-#include <omp.h>
 
 using namespace ImmGraphics;
 
@@ -16,24 +15,24 @@ public:
     virtual void StartPipeline(const VertexBuffer &vertices, const IndexBuffer &indices) override
     {
         #pragma omp parallel for
-        for (int i = 0; i < indices.getSize(); i += 3)
+        for (int i = 0; i < (int)indices.getSize(); i += 3)
         {
-            unsigned index0 = indices[i];
-            unsigned index1 = indices[i + 1];
-            unsigned index2 = indices[i + 2];
+            size_t index0 = indices[static_cast<size_t>(i)];
+            size_t index1 = indices[static_cast<size_t>(i) + 1];
+            size_t index2 = indices[static_cast<size_t>(i) + 2];
 
             Vertex v0 = vertices[index0];
             Vertex v1 = vertices[index1];
             Vertex v2 = vertices[index2];
 
-            Matrix4 transform = Matrix4::Viewport(m_device->width, m_device->height);
+            Matrix4 transform = Matrix4::Viewport((float)m_device->width, (float)m_device->height);
             v0.pos = transform * Vec4(v0.pos, 1);
             v1.pos = transform * Vec4(v1.pos, 1);
             v2.pos = transform * Vec4(v2.pos, 1);
 
-            setLine(v0.pos.x, v0.pos.y, v1.pos.x, v1.pos.y, v0.color);
-            setLine(v1.pos.x, v1.pos.y, v2.pos.x, v2.pos.y, v1.color);
-            setLine(v2.pos.x, v2.pos.y, v0.pos.x, v0.pos.y, v2.color);
+            setLine((int)v0.pos.x, (int)v0.pos.y, (int)v1.pos.x, (int)v1.pos.y, v0.color);
+            setLine((int)v1.pos.x, (int)v1.pos.y, (int)v2.pos.x, (int)v2.pos.y, v1.color);
+            setLine((int)v2.pos.x, (int)v2.pos.y, (int)v0.pos.x, (int)v0.pos.y, v2.color);
 
         }
     }
